requirements.txt
streamlit
python-docx